Redirect to main repository website
I initially published the updated documentation here and encouraged people to link to it until the library maintainers got around to reviewing the pull request. Now that the Doxygen documentation is published on the main repository, I'm changing my forked GitHub Pages site to a simple redirect to encourage others to use the official documentation.